requirements.txt
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
fonttools==4.34.4
idna==3.3
install==1.3.5
joblib==1.1.0
jycm==1.3.0
kiwisolver==1.4.4
matplotlib==3.5.2
modAL-python==0.4.2.1
numpy==1.23.1
packaging==21.3
panda==0.3.1
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.9.1
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
skorch==0.9.0
tabulate==0.8.10
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
